modified db structure

+++ data/sql/postgres/common/tables/dimensions.sql
@@ -2,7 +2,7 @@
  * dimensions table
  */
 create table if not exists dimensions (
-    id tinyint not null,
+    id smallint not null,
     name varchar(20) not null,
     constraint pk_dimensions primary key (id),
     constraint u_dimensions_1 unique (name)
